+++ docs/history.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Values for environment `env-vars` now support context formatting
 - Add `name` override for environments to allow for regular expression matching
 - The `index` publisher now better supports non-PyPI indices
+- Add certificate options to the `index` publisher
 - Display help text the first time the `shell` command is executed
 - Add support for Almquist (`ash`) shells
 - Add `hyperlink` as a dependency for better handling of package index URLs